Add files via upload
styling the web page for mobile phone. They key component i want to highlight is the navigation panel. its a container that has the nav element. i gave it no width , or height, but these are fetures that should be considered. the padding is excluded to allow complete disappearance when the max height is set to 0, so we dont have to add an additional nullifying instruction.